requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyjwt==2.9.0
argon2-cffi==23.1.0
pydantic==2.9.2